+++ src/main/java/de/dafuqs/spectrum/blocks/enchanter/EnchanterBlockEntity.java
@@ -120,7 +120,7 @@ public class EnchanterBlockEntity extends InWorldInteractionBlockEntity implemen
 			world.addParticle(ColoredCraftingParticleEffect.LIME, blockPos.getX() + randomX, blockPos.getY() + 2.5 + randomY, blockPos.getZ() + randomZ, 0.0D, -0.1D, 0.0D);
 			
 			if (world.getGameTime() % 12 == 0) {
-				world.playSound(null, enchanterBlockEntity.worldPosition, SoundEvents.EXPERIENCE_ORB_PICKUP, SoundSource.BLOCKS, 0.8F * SpectrumConfig.CONFIG.BlockSoundVolume.get(), 0.8F + world.random.nextFloat() * 0.4F);
+				world.playSound(null, enchanterBlockEntity.worldPosition, SoundEvents.EXPERIENCE_ORB_PICKUP, SoundSource.BLOCKS, (float) (0.8D * SpectrumConfig.CONFIG.BlockSoundVolume.get()), 0.8F + world.random.nextFloat() * 0.4F);
 				enchanterBlockEntity.doItemBowlOrbs(world);
 			}
 		}
